server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.7.0",
    "dj-database-url>=2.2.0",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "google-genai>=1.66.0",
    "langchain[google-genai]>=1.2.10",
    "langsmith>=0.7.16",
    "pillow>=12.1.1",
    "python-dotenv>=1.1.1",
    "pywebpush>=2.0.3",
    "psycopg[binary]>=3.2.0",
    "whitenoise>=6.6.0",
    "pyjwt>=2.12.1",
    "cryptography>=46.0.5",
]
